solvault/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pydantic==2.7.1
supabase>=2.10.0
google-generativeai==0.7.2
solders>=0.21.0
solana>=0.34.0
requests>=2.31.0
